Refactored the stream pull-parsing esque features out of the UBJInputStream and into its own sub class abstraction.
This way for RAW serial/deserial, the UBJInput/Output streams are
synonymous and work in the expected manner.

For a more "pull parser" experience, the UBJInputStreamParser will
provide the ability to "peek" at the next entity marker so you can then
call the appropriate readXXX method to parse it in a switch block.

This also made UBJInputStream faster.