Merged implementation of InputStreamParser and plain InputStream.
After going to implement ObjectReader, it occurred to me that the
UBJInputStream, in order to be realistically useful to anyone, had to
implement the core parser-style behavior of the InputStreamParser.

Looking through the code, there was a lot of ugly duplication and
shallow overriding in UBJInputStreamParser just to accomplish this
mental separation, but with a trivial amount of work this functionality
(Without the class or processing overhead) could be merged directly into
the core UBJInputStream.