Performance!
Massive performance boost in rewrite of StreamDecoder. Removed the need
for a write buffer entirely AND removed both the full char[] copy before
returning the result as well as the potential error condition when
char[] would have to be recopied when working with high-byte UTF-8 chars
where the number of bytes is greater than the number of chars
represented.

Resulting a raw 20-25% increase in performance.