Much more optimized implementation of decode.
ByteBuffer is no longer re-created on every single call; the single
instance wrapping the byte[] buffer is used and re-used every single
time.

Caller must provide a CharBuffer big enough to hold the number of chars
being decoded; previously this was being created inside of the decode
method, but requiring the caller to send one (re-use) allows them to
realize the requirements of this method and possibly cache/re-use
buffers as many times as they want to save on the overhead.

This method is *raw* decode logic now with next to no object creation
and GC except what the underlying CharsetDecoder creates and uses.